requirements.txt
tensorflow
fasttext
sklearn
faiss-cpu